Bring back service call callback to Rx API
