+++ src/App/NetDaemon.App/Common/NetDaemonAppBase.cs
@@ -22,6 +22,17 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Common
         private Task? _manageRuntimeInformationUpdatesTask;
 
         /// <summary>
+        ///     Registered callbacks for service calls
+        /// </summary>
+        private readonly List<(string, string, Func<dynamic?, Task>)> _daemonCallBacksForServiceCalls
+            = new List<(string, string, Func<dynamic?, Task>)>();
+
+        /// <summary>
+        ///     All actions being performed for service call events
+        /// </summary>
+        public List<(string, string, Func<dynamic?, Task>)> DaemonCallBacksForServiceCalls => _daemonCallBacksForServiceCalls;
+
+        /// <summary>
         ///     Next scheduled time
         /// </summary>
         protected DateTime? NextScheduledEvent { get; set; } = null;
@@ -204,6 +215,8 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Common
             if (_manageRuntimeInformationUpdatesTask is object)
                 await _manageRuntimeInformationUpdatesTask.ConfigureAwait(false);
 
+            _daemonCallBacksForServiceCalls.Clear();
+            
             this.IsEnabled = false;
             _lazyStoreStateTask = null;
             _storageObject = null;
@@ -217,6 +230,10 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Common
             _ = _daemon as INetDaemon ?? throw new NullReferenceException($"{nameof(_daemon)} cant be null!");
             return _daemon!.GetApp(appInstanceId);
         }
+
+        /// <inheritdoc/>
+        public void ListenServiceCall(string domain, string service, Func<dynamic?, Task> action)
+            => _daemonCallBacksForServiceCalls.Add((domain.ToLowerInvariant(), service.ToLowerInvariant(), action));
 
         /// <inheritdoc/>
         public void Log(string message) => Log(LogLevel.Information, message);
