+++ src/Daemon/NetDaemon.Daemon/Daemon/DaemonAppExtensions.cs
@@ -1,4 +1,5 @@
 using JoySoftware.HomeAssistant.NetDaemon.Common;
+using JoySoftware.HomeAssistant.NetDaemon.Common.Reactive;
 using JoySoftware.HomeAssistant.NetDaemon.Daemon.Config;
 using Microsoft.Extensions.Logging;
 using System;
@@ -26,7 +27,7 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Daemon
                         switch (attr)
                         {
                             case HomeAssistantServiceCallAttribute hasstServiceCallAttribute:
-                                await HandleServiceCallAttribute(_daemon, daemonApp, method).ConfigureAwait(false);
+                                await HandleServiceCallAttribute(_daemon, daemonApp, method, true).ConfigureAwait(false);
                                 break;
 
                             case HomeAssistantStateChangedAttribute hassStateChangedAttribute:
@@ -34,13 +35,22 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Daemon
                                 break;
                         }
                     }
+                    else if (netDaemonApp is NetDaemonRxApp daemonRxApp)
+                    {
+                        switch (attr)
+                        {
+                            case HomeAssistantServiceCallAttribute hasstServiceCallAttribute:
+                                await HandleServiceCallAttribute(_daemon, daemonRxApp, method, false).ConfigureAwait(false);
+                                break;
+                        }
+                    }
                 }
             }
         }
 
-        private static (bool, string) CheckIfServiceCallSignatureIsOk(MethodInfo method)
+        private static (bool, string) CheckIfServiceCallSignatureIsOk(MethodInfo method, bool async)
         {
-            if (method.ReturnType != typeof(Task))
+            if (async && method.ReturnType != typeof(Task))
                 return (false, $"{method.Name} has not correct return type, expected Task");
 
             var parameters = method.GetParameters();
@@ -78,9 +88,9 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Daemon
             return (true, string.Empty);
         }
 
-        private static async Task HandleServiceCallAttribute(INetDaemon _daemon, NetDaemonApp netDaemonApp, MethodInfo method)
+        private static async Task HandleServiceCallAttribute(INetDaemon _daemon, NetDaemonAppBase netDaemonApp, MethodInfo method, bool async=true)
         {
-            var (signatureOk, err) = CheckIfServiceCallSignatureIsOk(method);
+            var (signatureOk, err) = CheckIfServiceCallSignatureIsOk(method, async);
             if (!signatureOk)
             {
                 _daemon.Logger.LogWarning(err);
@@ -98,6 +108,7 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Daemon
                     try
                     {
                         var expObject = data as ExpandoObject;
+
                         await method.InvokeAsync(netDaemonApp, expObject!).ConfigureAwait(false);
                     }
                     catch (Exception e)
@@ -106,6 +117,7 @@ namespace JoySoftware.HomeAssistant.NetDaemon.Daemon
                     }
                 });
         }
+        
 
         private static void HandleStateChangedAttribute(
                                     INetDaemon _daemon,
